% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glioGSC}
\alias{glioGSC}
\title{msigdb: 47 gene sets related to glioblastoma by 'Search Gene Sets' at msigdb}
\format{GeneSetCollection}
\usage{
glioGSC
}
\description{
msigdb: 47 gene sets related to glioblastoma by 'Search Gene Sets' at msigdb
}
\note{
Retrieved 20 Oct 2017, and imported with getGmt of GSEABase
}
\keyword{datasets}
