% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivyGlimpse.R
\name{ivyGlimpse}
\alias{ivyGlimpse}
\title{simple app to explore image property quantifications in relation to survival and expression}
\usage{
ivyGlimpse()
}
\description{
simple app to explore image property quantifications in relation to survival and expression
}
