% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ivySE}
\alias{ivySE}
\title{ivySE: SummarizedExperiment for IvyGAP expression data and metadata}
\format{SummarizedExperiment instance}
\source{
processed from \url{glioblastoma.alleninstitute.org}
}
\usage{
ivySE
}
\description{
ivySE: SummarizedExperiment for IvyGAP expression data and metadata
}
\keyword{datasets}
