% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets.R
\name{makeGeneSets}
\alias{makeGeneSets}
\title{demonstration of gene set construction for ivyGlimpse app}
\usage{
makeGeneSets()
}
\value{
list of gene sets
}
\description{
demonstration of gene set construction for ivyGlimpse app
}
\note{
Should be replaced by selections from a general catalogy
}
\examples{
str(makeGeneSets())
}
